Global Styles: Add colors and typograpghy to the browse styles section
